+++ burner/extension/src/lib/inpage/message.ts
@@ -1,5 +1,3 @@
-import { account } from "./account";
-
 const uuid = "589c80c1eb85413d";
 
 export type burnerMessage = {
@@ -8,20 +6,27 @@ export type burnerMessage = {
   data: any;
 };
 
-export const request = (msg: burnerMessage) => {
-  if (!msg || typeof msg !== "object" || !msg.type) {
-    throw new Error("Invalid message");
-  }
-  const burner = document.querySelector("#starknetburner");
-  if (!burner) {
-    throw new Error("wallet not found");
-  }
-  const iframe = burner.querySelector<HTMLIFrameElement>("#iframe");
-  if (!iframe) {
-    throw new Error("iframe not found");
-  }
-  console.log("sending now...", { ...msg, uuid });
-  return iframe.contentWindow?.postMessage({ ...msg, uuid }, "*");
+export const requestFactory = (
+  debug: boolean
+): ((msg: burnerMessage) => void) => {
+  const request = (msg: burnerMessage) => {
+    if (!msg || typeof msg !== "object" || !msg.type) {
+      throw new Error("Invalid message");
+    }
+    const burner = document.querySelector("#starknetburner");
+    if (!burner) {
+      throw new Error("wallet not found");
+    }
+    const iframe = burner.querySelector<HTMLIFrameElement>("#iframe");
+    if (!iframe) {
+      throw new Error("iframe not found");
+    }
+    if (debug) {
+      console.log("sending now...", { ...msg, uuid });
+    }
+    return iframe.contentWindow?.postMessage({ ...msg, uuid }, "*");
+  };
+  return request;
 };
 
 export const waitForMessage = (
@@ -41,11 +46,17 @@ export const waitForMessage = (
   });
 };
 
-export const extensionEventHandler = (messageEvent: MessageEvent) => {
-  if (messageEvent?.data?.uuid === uuid) {
-    account._events[messageEvent.data.type].forEach((fn: any) => {
-      console.log(`event ${messageEvent.data.type} received in extension`);
-      fn(messageEvent.data.data);
-    });
+export const extensionEventHandler = (event: MessageEvent) => {
+  if (event?.data?.uuid !== uuid) {
+    return;
+  }
+  const { type, data } = event.data;
+  switch (type) {
+    case "pong":
+      console.log("in:extension", type, data);
+      break;
+    default:
+      console.log("in:extension", "default event", type);
+      break;
   }
 };
