+++ burner/extension/src/lib/inpage/signer.ts
